chemistry/requirements.txt
black==22.3.0
boto3==1.26.38
chardet==4.0.0
charset-normalizer==2.0.12
Django==3.2.16
django-cors-headers==3.11.0
django-csvimport==2.16
django-extensions==3.1.5
django-extra-fields==3.0.2
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-csv==2.1.1
django-environ==0.9.0
drf-spectacular==0.24.0
flake8==4.0.1
flake8-black==0.3.2
haversine==2.7.0
isort==5.10.1
nodeenv==1.6.0
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
pylint==2.13.7
pyparsing==3.0.8
pyrsistent==0.18.1
python-dateutil==2.8.2
python-decouple==3.6
python-redis-lock==3.7.0
psycopg2-binary==2.9.3
python-dotenv==0.21.0
PyYAML==6.0
redis==4.2.2
regex==2022.3.15
requests==2.27.1
urllib3==1.26.9
ssm-parameter-store==19.11.0
typing_extensions==4.2.0
vine==5.0.0
virtualenv==20.14.1
